Polish failure summary
This polishes the failure summary by grouping and sorting errors by the keys
they originated from. Incidentally, this fixes an issue with the keys for errors
of Map derivations not being properly concatenated, and makes the output of the
WrongType failure more uniform, by always referring to the appropriate
ConfigValueTypes.