Fix broken and redirecting links (#438)
Fix broken links mentioned by @jcranky in gitter. 

Also, generally replace typesafehub with lightbend in URLs to avoid redirects.

Also, don't allow htmlproofer to validate links on search.maven.org, which is what maven-badges redirects to, because it does a diabolical thing to trick spiders: it returns a 404 for every page with a body that uses JS to load the actual page.