Add methods to derive enumerations
Deprecate EnumCoproductHint in favor of methods in the `semiauto` package that
allow the derivation of readers, writers and full converters for enumerations
encoded as sealed traits. In a semi-automatic setting, this allows the
derivation of enumeration-based type classes for the sealed trait type without
requiring implicit type classes in scope for all the subtypes.