Support for Scala 2.13 in cron4s
`cron4s` supports Scala 2.13 in the 0.6 version series, but at the same time drops support for Scala 2.11. This change makes this module available for all the cross Scala versions at the cost of choosing a different version of the dependency.

Other option would be dropping 2.11 support in this module.

ps: `cron4s` 0.6 will be source compatible with 0.5.