Improve support for more exotic ConfigValueLocation
Hey,

I've noticed that pureconfig doesn't print a useful error location when you parse a String with `ConfigFactory.parseString`. In the `ConfigParseOptions` you can use `setOriginDescription` to give the user a hint where the config is from, but that won't set a url for the `ConfigOrigin`, and hence the current implementation in pureconfig won't use it.

A better way to do it is to use `ConfigOrigin`'s `description` method because that will return something useful for all origins, whether it's from a file, a resource etc.. Since it never returns null, no null check is needed.