Generate `Automatic-Module-Name` attribute
It is currently impossible to define `pureconfig` on a module path. This
is because Java fails to automatically infer Module name from jar's
containing invalid characters (e.g. `_` or `-` in `_2.13` cross version).
It's a common problem with Scala artifacts.

This change adds `Automatic-Module-Name` attribute to MANIFEST that
contains a valid Java identifier that can be used to define appropriate
module dependencies.